# swaggerstrayV3.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Stray",
		"contact": {},
		"version": "1.0"
	},
	"servers": [
		{
			"url": "http://localhost:3000/api/v1",
			"variables": {}
		}
	],
	"paths": {
...
		"/member/items": {
			"get": {
				"tags": [
					"Items"
				],
				"summary": "getItems",
				"description": "getItems",
				"operationId": "getItems",
				"parameters": [
					{
						"name": "Authorization",
						"in": "header",
						"description": "",
						"style": "simple",
						"schema": {
							"type": "string",
							"default": "Bearer {token}"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {}
					}
				},
				"deprecated": false
			}
		},
...
		"/member/carts": {
			"post": {
				"tags": [
					"Cart"
				],
				"summary": "createCart",
				"description": "createCart",
				"operationId": "createCart",
				"parameters": [
					{
						"name": "Authorization",
						"in": "header",
						"description": "",
						"style": "simple",
						"schema": {
							"type": "string",
							"default": "Bearer {token}"
						}
					}
				],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"allOf": [
									{
										"$ref": "#/components/schemas/createCartRequest"
									},
									{
										"example": {
											"id_item": 1,
											"quantity": 1
										}
									}
								]
							},
							"example": {
								"id_item": 1,
								"quantity": 1
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {}
					}
				},
				"deprecated": false
			},
			"get": {
				"tags": [
					"Cart"
				],
				"summary": "getAllCart",
				"description": "getAllCart",
				"operationId": "getAllCart",
				"parameters": [
					{
						"name": "Authorization",
						"in": "header",
						"description": "",
						"style": "simple",
						"schema": {
							"type": "string",
							"default": "Bearer {token}"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {}
					}
				},
				"deprecated": false
			}
		},
		"/member/carts/3": {
			"patch": {
				"tags": [
					"Cart"
				],
				"summary": "updateCart",
				"description": "updateCart",
				"operationId": "updateCart",
				"parameters": [
					{
						"name": "Authorization",
						"in": "header",
						"description": "",
						"style": "simple",
						"schema": {
							"type": "string",
							"default": "Bearer {token}"
						}
					}
				],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"allOf": [
									{
										"$ref": "#/components/schemas/updateCartRequest"
									},
									{
										"example": {
											"quantity": 2
										}
									}
								]
							},
							"example": {
								"quantity": 2
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {}
					}
				},
				"deprecated": false
			},
			"delete": {
				"tags": [
					"Cart"
				],
				"summary": "deleteCart",
				"description": "deleteCart",
				"operationId": "deleteCart",
				"parameters": [
					{
						"name": "Authorization",
						"in": "header",
						"description": "",
						"style": "simple",
						"schema": {
							"type": "string",
							"default": "Bearer {token}"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {}
					}
				},
				"deprecated": false
			}
		},
		"/auth/refresh_token": {
			"post": {
				"tags": [
					"Misc"
				],
				"summary": "RefreshToken",
				"description": "Refresh Token",
				"operationId": "RefreshToken",
				"parameters": [
					{
						"name": "Authorization",
						"in": "header",
						"description": "",
						"style": "simple",
						"schema": {
							"type": "string",
							"default": "Bearer {token}"
						}
					}
				],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"allOf": [
									{
										"$ref": "#/components/schemas/RefreshTokenRequest"
									},
									{
										"example": {
											"email": "akhmad.maknur@gmail.com",
											"token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImFraG1hZC5tYWtudXJAZ21haWwxLmNvbSIsInVzZXJfaWQiOjIsImlhdCI6MTY2MDg4Nzk1MiwiZXhwIjoxNjYwODkxNTUyfQ.EvFi8la6yv84YsENWRZQPvJMMA8zfrseGj7B8FLuD0I"
										}
									}
								]
							},
							"example": {
								"email": "akhmad.maknur@gmail.com",
								"token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImFraG1hZC5tYWtudXJAZ21haWwxLmNvbSIsInVzZXJfaWQiOjIsImlhdCI6MTY2MDg4Nzk1MiwiZXhwIjoxNjYwODkxNTUyfQ.EvFi8la6yv84YsENWRZQPvJMMA8zfrseGj7B8FLuD0I"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {}
					}
				},
				"deprecated": false
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"createCartRequest": {
				"title": "createCartRequest",
				"required": [
					"id_item",
					"quantity"
				],
				"type": "object",
				"properties": {
					"id_item": {
						"type": "integer",
						"format": "int32"
					},
					"quantity": {
						"type": "integer",
						"format": "int32"
					}
				},
				"example": {
					"id_item": 1,
					"quantity": 1
				}
			},
			"updateCartRequest": {
				"title": "updateCartRequest",
				"required": [
					"quantity"
				],
				"type": "object",
				"properties": {
					"quantity": {
						"type": "integer",
						"format": "int32"
					}
				},
				"example": {
					"quantity": 2
				}
			},
			"RefreshTokenRequest": {
				"title": "RefreshTokenRequest",
				"required": [
					"email",
					"token"
				],
				"type": "object",
				"properties": {
					"email": {
						"type": "string"
					},
					"token": {
						"type": "string"
					}
				},
				"example": {
					"email": "akhmad.maknur@gmail.com",
					"token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImFraG1hZC5tYWtudXJAZ21haWwxLmNvbSIsInVzZXJfaWQiOjIsImlhdCI6MTY2MDg4Nzk1MiwiZXhwIjoxNjYwODkxNTUyfQ.EvFi8la6yv84YsENWRZQPvJMMA8zfrseGj7B8FLuD0I"
				}
			},
...
		}
	},
...
}